Move LdapUser
Move to the domain.ldap package
